LayoutTests/fast/canvas/webgl/uniform-location-expected.txt

Tests WebGL APIs related to shader uniforms

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS contextA.useProgram(programA2) is undefined.
PASS contextA.uniformMatrix4fv(locationA, false, mat) threw exception GL error 1282 in uniformMatrix4fv.
PASS contextA.useProgram(programA1) is undefined.
PASS contextA.uniformMatrix4fv(locationA, false, mat) is undefined.
PASS contextA.uniformMatrix4fv(0, false, mat) threw exception GL error 1281 in uniformMatrix4fv.
PASS contextA.useProgram(programS) is undefined.
PASS contextA.uniform1i(locationSx, 3) is undefined.
PASS contextA.uniform1f(locationArray0, 4.0) is undefined.
PASS contextA.getUniform(programS, locationSx) is 3
PASS contextA.getUniform(programS, locationArray0) is 4.0
PASS contextA.useProgram(programV) is undefined.
PASS contextA.uniform4fv(locationVec4, vec) is undefined.
PASS contextA.getUniform(programV, locationVec4) is vec
PASS contextA.getUniformLocation(programV, "IDontExist") is null
PASS successfullyParsed is true

TEST COMPLETE
